Add test for MergeInputReader in HandlerTestBase, check if handler was called at all
